+++ src/main/java/org/opentripplanner/datastore/configure/DataStoreModule.java
@@ -6,10 +6,10 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import javax.inject.Singleton;
-import org.opentripplanner.datastore.CompositeDataSource;
-import org.opentripplanner.datastore.FileType;
 import org.opentripplanner.datastore.OtpDataStore;
-import org.opentripplanner.datastore.OtpDataStoreConfig;
+import org.opentripplanner.datastore.api.CompositeDataSource;
+import org.opentripplanner.datastore.api.FileType;
+import org.opentripplanner.datastore.api.OtpDataStoreConfig;
 import org.opentripplanner.datastore.base.DataSourceRepository;
 import org.opentripplanner.datastore.file.FileDataSourceRepository;
 import org.opentripplanner.ext.datastore.gs.GsDataSourceRepository;
