+++ src/main/java/org/opentripplanner/graph_builder/model/GtfsBundle.java
@@ -4,8 +4,8 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import org.onebusaway.csv_entities.CsvInputSource;
-import org.opentripplanner.datastore.CompositeDataSource;
-import org.opentripplanner.datastore.FileType;
+import org.opentripplanner.datastore.api.CompositeDataSource;
+import org.opentripplanner.datastore.api.FileType;
 import org.opentripplanner.datastore.configure.DataStoreModule;
 import org.opentripplanner.graph_builder.module.GtfsFeedId;
 import org.slf4j.Logger;
