+++ src/main/java/org/opentripplanner/netex/NetexBundle.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 import javax.xml.bind.JAXBException;
-import org.opentripplanner.datastore.CompositeDataSource;
-import org.opentripplanner.datastore.DataSource;
+import org.opentripplanner.datastore.api.CompositeDataSource;
+import org.opentripplanner.datastore.api.DataSource;
 import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.model.impl.OtpTransitServiceBuilder;
 import org.opentripplanner.netex.index.NetexEntityIndex;
