+++ src/main/java/org/opentripplanner/netex/loader/NetexDataSourceHierarchy.java
@@ -6,8 +6,8 @@ import java.util.Map;
 import java.util.TreeMap;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import org.opentripplanner.datastore.CompositeDataSource;
-import org.opentripplanner.datastore.DataSource;
+import org.opentripplanner.datastore.api.CompositeDataSource;
+import org.opentripplanner.datastore.api.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
