+++ src/main/java/org/opentripplanner/openstreetmap/OpenStreetMapProvider.java
@@ -6,8 +6,8 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import org.openstreetmap.osmosis.osmbinary.file.BlockInputStream;
-import org.opentripplanner.datastore.DataSource;
-import org.opentripplanner.datastore.FileType;
+import org.opentripplanner.datastore.api.DataSource;
+import org.opentripplanner.datastore.api.FileType;
 import org.opentripplanner.datastore.file.FileDataSource;
 import org.opentripplanner.graph_builder.module.osm.OSMDatabase;
 import org.opentripplanner.util.logging.ProgressTracker;
