+++ src/main/java/org/opentripplanner/standalone/configure/OTPApplicationModule.java
@@ -3,7 +3,7 @@ package org.opentripplanner.standalone.configure;
 import dagger.Module;
 import dagger.Provides;
 import javax.inject.Singleton;
-import org.opentripplanner.datastore.OtpDataStoreConfig;
+import org.opentripplanner.datastore.api.OtpDataStoreConfig;
 import org.opentripplanner.standalone.config.BuildConfig;
 import org.opentripplanner.standalone.config.CommandLineParameters;
 import org.opentripplanner.standalone.config.OtpConfig;
